% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GradDesc.R
\docType{package}
\name{GradDesc-package}
\alias{GradDesc-package}
\title{GradDesc}
\description{
GradDesc-package: paquete wrapper de la biblioteca Purple
}
\details{
La documentacion de la clase \code{GradientDescentSumaCuadradosClase} está en: \code{\link{GradientDescentSumaCuadradosClase}}
}
\author{
Gonzalo Mateo, Paulino Tardáguila
}

